local repo migration

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: agarbacz <agarbacz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 13:05:17 by agarbacz          #+#    #+#             */
-/*   Updated: 2024/11/29 14:57:35 by agarbacz         ###   ########.fr       */
+/*   Updated: 2024/12/04 15:34:31 by agarbacz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
